ORDERING STEPS! BACKEND SIDE

+++ api/database/migrations/2020_06_07_175955_create_steps_table.php
@@ -20,7 +20,7 @@ class CreateStepsTable extends Migration
                 ->onDelete('cascade');
             $table->uuid('uuid');
             $table->string('title')->nullable();
-            $table->unsignedInteger('order')->index();
+            $table->decimal('order')->index();
             $table->string('body')->nullable();
 
             $table->timestamps();
